+++ resources/js/visual-editor/blocks/columns/index.ts
@@ -15,6 +15,7 @@ import variations from './variations';
 import icon from './inserter-icon';
 
 import './columns.css';
+import '../../../../css/flex-layout.css';
 
 export {
     edit,